+++ tests/conftest.py
@@ -16,7 +16,7 @@ sys.path.insert(0, _ROOT)
 
 JWT_PUBLIC_KEY = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC8LrEp0Q6l1WPsY32uOPqEjaisQScnzO/XvlhQTzj5w+hFObjiNgIaHRceYh3hZZwsRsHIkCxOY0JgUPeFP9IVXso0VptIjCPRF5yrV/+dF1rtl4eyYj/XOBvSDzbQQwqdjhHffw0TXW0f/yjGGJCYM+tw/9dmj9VilAMNTx1H76uPKUo4M3vLBQLo2tj7z1jlh4Jlw5hKBRcWQWbpWP95p71Db6gSpqReDYbx57BW19APMVketUYsXfXTztM/HWz35J9HDya3ID0Dl+pE22Wo8SZo2+ULKu/4OYVcD8DjF15WwXrcuFDypX132j+LUWOVWxCs5hdMybSDwF3ZhVBH ec2-user@ip-172-31-41-191.eu-west-1.compute.internal"  # NOQA
 
-MONGO_URI = os.environ.get("MONGO_URI", "mongodb://localhost:27017/")
+TEST_MONGO_URI = os.environ.get("TEST_MONGO_URI", "mongodb://localhost:27017/")
 
 
 def random_string(
@@ -94,10 +94,10 @@ def tmp_app(request):
         },
         "SECRET_KEY": "secret",
         "SQLALCHEMY_DATABASE_URI": "sqlite:///:memory:",
-        "RETRIEVE_MONGO_URI": MONGO_URI,
+        "RETRIEVE_MONGO_URI": TEST_MONGO_URI,
         "RETRIEVE_MONGO_DB": tmp_mongo_db_name,
         "RETRIEVE_MONGO_COLLECTION": "datasets",
-        "SEARCH_MONGO_URI": MONGO_URI,
+        "SEARCH_MONGO_URI": TEST_MONGO_URI,
         "SEARCH_MONGO_DB": tmp_mongo_db_name,
         "SEARCH_MONGO_COLLECTION": "datasets",
         "SQLALCHEMY_TRACK_MODIFICATIONS": False,
@@ -106,6 +106,9 @@ def tmp_app(request):
         "JWT_TOKEN_LOCATION": "headers",
         "JWT_HEADER_NAME": "Authorization",
         "JWT_HEADER_TYPE": "Bearer",
+        "MONGO_URI": os.environ.get("TEST_MONGO_URI", "mongodb://localhost:27017/"),
+        "MONGO_DB": tmp_mongo_db_name,
+        "MONGO_COLLECTION": "dependencies",
     }
 
     app = create_app(config)
@@ -153,10 +156,10 @@ def tmp_app_with_users(request):
         "OPENAPI_VERSION": '3.0.2',
         "SECRET_KEY": "secret",
         "SQLALCHEMY_DATABASE_URI": "sqlite:///:memory:",
-        "RETRIEVE_MONGO_URI": MONGO_URI,
+        "RETRIEVE_MONGO_URI": TEST_MONGO_URI,
         "RETRIEVE_MONGO_DB": tmp_mongo_db_name,
         "RETRIEVE_MONGO_COLLECTION": "datasets",
-        "SEARCH_MONGO_URI": MONGO_URI,
+        "SEARCH_MONGO_URI": TEST_MONGO_URI,
         "SEARCH_MONGO_DB": tmp_mongo_db_name,
         "SEARCH_MONGO_COLLECTION": "datasets",
         "SQLALCHEMY_TRACK_MODIFICATIONS": False,
@@ -165,6 +168,9 @@ def tmp_app_with_users(request):
         "JWT_TOKEN_LOCATION": "headers",
         "JWT_HEADER_NAME": "Authorization",
         "JWT_HEADER_TYPE": "Bearer",
+        "MONGO_URI": os.environ.get("TEST_MONGO_URI", "mongodb://localhost:27017/"),
+        "MONGO_DB": tmp_mongo_db_name,
+        "MONGO_COLLECTION": "dependencies",
     }
 
     app = create_app(config)
@@ -227,10 +233,10 @@ def tmp_app_with_data(request):
         "API_VERSION": 'v1',
         "OPENAPI_VERSION": '3.0.2',
         "SQLALCHEMY_DATABASE_URI": "sqlite:///:memory:",
-        "RETRIEVE_MONGO_URI": MONGO_URI,
+        "RETRIEVE_MONGO_URI": TEST_MONGO_URI,
         "RETRIEVE_MONGO_DB": tmp_mongo_db_name,
         "RETRIEVE_MONGO_COLLECTION": "datasets",
-        "SEARCH_MONGO_URI": MONGO_URI,
+        "SEARCH_MONGO_URI": TEST_MONGO_URI,
         "SEARCH_MONGO_DB": tmp_mongo_db_name,
         "SEARCH_MONGO_COLLECTION": "datasets",
         "SQLALCHEMY_TRACK_MODIFICATIONS": False,
@@ -239,6 +245,9 @@ def tmp_app_with_data(request):
         "JWT_TOKEN_LOCATION": "headers",
         "JWT_HEADER_NAME": "Authorization",
         "JWT_HEADER_TYPE": "Bearer",
+        "MONGO_URI": os.environ.get("TEST_MONGO_URI", "mongodb://localhost:27017/"),
+        "MONGO_DB": tmp_mongo_db_name,
+        "MONGO_COLLECTION": "dependencies",
     }
 
     app = create_app(config)
@@ -352,14 +361,17 @@ def tmp_cli_runner(request):
         "API_VERSION": 'v1',
         "OPENAPI_VERSION": '3.0.2',
         "SQLALCHEMY_DATABASE_URI": "sqlite:///:memory:",
-        "RETRIEVE_MONGO_URI": MONGO_URI,
+        "RETRIEVE_MONGO_URI": TEST_MONGO_URI,
         "RETRIEVE_MONGO_DB": tmp_mongo_db_name,
         "RETRIEVE_MONGO_COLLECTION": "datasets",
-        "SEARCH_MONGO_URI": MONGO_URI,
+        "SEARCH_MONGO_URI": TEST_MONGO_URI,
         "SEARCH_MONGO_DB": tmp_mongo_db_name,
         "SEARCH_MONGO_COLLECTION": "datasets",
         "SQLALCHEMY_TRACK_MODIFICATIONS": False,
-        "SECRET_KEY": "dev"
+        "SECRET_KEY": "dev",
+        "MONGO_URI": os.environ.get("TEST_MONGO_URI", "mongodb://localhost:27017/"),
+        "MONGO_DB": tmp_mongo_db_name,
+        "MONGO_COLLECTION": "dependencies",
     }
 
     app = create_app(config)