+++ tests/test_me_routes.py
@@ -22,7 +22,8 @@ def test_me_route(
             'is_admin': True,
             'register_permissions_on_base_uris': [],
             'search_permissions_on_base_uris': [],
-            'username': 'snow-white'
+            'username': 'snow-white',
+            'display_name': None
         })
 
     r = tmp_app_with_users_client.get(
@@ -45,7 +46,8 @@ def test_me_route(
             'is_admin': False,
             'register_permissions_on_base_uris': ['s3://snow-white'],
             'search_permissions_on_base_uris': ['s3://snow-white'],
-            'username': 'grumpy'
+            'username': 'grumpy',
+            'display_name': None
         })
 
     headers = dict(Authorization="Bearer " + grumpy_token)
@@ -102,7 +104,10 @@ def test_me_summary_route(
         "size_in_bytes_per_base_uri": {},
         "tags": [],
         "datasets_per_tag": {},
-        "size_in_bytes_per_tag": {}
+        "size_in_bytes_per_tag": {},
+        "uploaders": [],
+        "datasets_per_uploader": {},
+        "size_in_bytes_per_uploader": {},
     }
     assert expected_content == json.loads(r.data.decode("utf-8"))
 
@@ -127,6 +132,9 @@ def test_me_summary_route(
         "tags": ["evil", "fruit", "good"],
         "datasets_per_tag": {"good": 1, "evil": 2, "fruit": 3},
         "size_in_bytes_per_tag": {"evil": 11483620, "fruit": 11483620, "good": 0},
+        "uploaders": [],
+        "datasets_per_uploader": {},
+        "size_in_bytes_per_uploader": {},
     }
     assert expected_content == json.loads(r.data.decode("utf-8"))
 
@@ -147,7 +155,10 @@ def test_me_summary_route(
         "size_in_bytes_per_base_uri": {},
         "tags": [],
         "datasets_per_tag": {},
-        "size_in_bytes_per_tag": {}
+        "size_in_bytes_per_tag": {},
+        "uploaders": [],
+        "datasets_per_uploader": {},
+        "size_in_bytes_per_uploader": {},
     }
     assert expected_content == json.loads(r.data.decode("utf-8"))